All integration tests moved to one directory.
