+++ src/test/scala-2.11/agents/integration/ABasicSubscriberPreCommitFailureTest.scala
@@ -1,4 +1,4 @@
-package agents.both
+package agents.integration
 
 import java.util.UUID
 import java.util.concurrent.locks.ReentrantLock
