+++ src/test/scala-2.11/agents/integration/ABasicSubscriberTotalAmountTest.scala
@@ -1,4 +1,4 @@
-package agents.both
+package agents.integration
 
 import java.util.UUID
 import java.util.concurrent.locks.ReentrantLock
@@ -7,18 +7,12 @@ import java.util.concurrent.{CountDownLatch, TimeUnit}
 import com.bwsw.tstreams.agents.consumer.Offset.Oldest
 import com.bwsw.tstreams.agents.consumer.subscriber.{Callback, SubscribingConsumer}
 import com.bwsw.tstreams.agents.producer.NewTransactionProducerPolicy
-import com.bwsw.tstreams.debug.GlobalHooks
 import com.bwsw.tstreams.env.TSF_Dictionary
 import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
-import testutils.{LocalGeneratorCreator, TestUtils}
+import testutils._
 
 //TODO refactoring
-class ABasicSubscriberAfterCommitFailureTest extends FlatSpec with Matchers
-  with BeforeAndAfterAll with TestUtils {
-
-  System.setProperty("DEBUG", "true")
-  GlobalHooks.addHook(GlobalHooks.afterCommitFailure, () => throw new RuntimeException)
-
+class ABasicSubscriberTotalAmountTest extends FlatSpec with Matchers with BeforeAndAfterAll with TestUtils {
   f.setProperty(TSF_Dictionary.Stream.NAME,"test_stream").
     setProperty(TSF_Dictionary.Stream.PARTITIONS,3).
     setProperty(TSF_Dictionary.Stream.TTL, 60 * 10).
@@ -33,7 +27,7 @@ class ABasicSubscriberAfterCommitFailureTest extends FlatSpec with Matchers
   val lock = new ReentrantLock()
   var acc = 0
   val totalTxns = 5
-  val dataInTxn = 1
+  val dataInTxn = 10
   val data = randomString
   val l1 = new CountDownLatch(1)
   val l2 = new CountDownLatch(1)
@@ -64,9 +58,8 @@ class ABasicSubscriberAfterCommitFailureTest extends FlatSpec with Matchers
       if (acc == totalTxns * 2)
         l2.countDown()
 
-      if (acc == totalTxns * 3) {
+      if (acc == totalTxns * 4)
         l3.countDown()
-      }
 
       lock.unlock()
     }
@@ -90,18 +83,34 @@ class ABasicSubscriberAfterCommitFailureTest extends FlatSpec with Matchers
     offset = Oldest,
     isUseLastOffset = true)
 
-  "subscribe consumer" should "retrieve all sent messages" in {
+  val s3 = f.getSubscriber[String](
+    name = "test_subscriber-2",
+    txnGenerator = LocalGeneratorCreator.getGen(),
+    converter = arrayByteToStringConverter,
+    partitions = List(0,1,2),
+    callback = callback,
+    offset = Oldest,
+    isUseLastOffset = true)
+
+  "Subscriber consumer with same name" should "retrieve all sent messages without duplicates" in {
 
     s1.start()
     sendTxnsAndWait(totalTxns, dataInTxn, data, l1)
-    sendTxnsAndWait(totalTxns, dataInTxn, data, l2)
     s1.stop()
 
     s2.start()
-    sendTxnsAndWait(totalTxns, dataInTxn, data, l3)
+    sendTxnsAndWait(totalTxns, dataInTxn, data, l2)
     s2.stop()
 
-    acc shouldEqual totalTxns * 3
+    acc shouldEqual totalTxns * 2
+  }
+
+  "Subscriber consumer with other name" should "retrieve all sent messages" in {
+    s3.start()
+    val r = l3.await(100000, TimeUnit.MILLISECONDS)
+    r shouldBe true
+    s3.stop()
+    acc shouldEqual totalTxns * 4
   }
 
   def sendTxnsAndWait(totalTxns: Int, dataInTxn: Int, data: String, l: CountDownLatch) = {
@@ -121,8 +130,6 @@ class ABasicSubscriberAfterCommitFailureTest extends FlatSpec with Matchers
   }
 
   override def afterAll(): Unit = {
-    System.clearProperty("DEBUG")
-    GlobalHooks.clear()
     producer.stop()
     onAfterAll()
   }
