+++ src/test/scala-2.11/agents/integration/GroupCheckpointTest.scala
@@ -1,4 +1,4 @@
-package agents.group
+package agents.integration
 
 import com.bwsw.tstreams.agents.consumer.Offset.Oldest
 import com.bwsw.tstreams.agents.group.CheckpointGroup
