+++ src/test/scala-2.11/agents/integration/ProducerEventReceiverTcpServerTests.scala
@@ -1,4 +1,4 @@
-package agents.both
+package agents.integration
 
 import com.bwsw.tstreams.coordination.messages.state.Message
 import com.bwsw.tstreams.coordination.subscriber.ProducerEventReceiverTcpServer
