+++ src/test/scala-2.11/agents/integration/ProducerUpdateTaskTest.scala
@@ -1,4 +1,4 @@
-package agents.producer
+package agents.integration
 
 import com.bwsw.tstreams.agents.producer.NewTransactionProducerPolicy
 import com.bwsw.tstreams.common.ResettableCountDownLatch
