+++ src/test/scala-2.11/agents/integration/ProducerWithManyOpenedTxnsTest.scala
@@ -1,4 +1,4 @@
-package agents.producer
+package agents.integration
 
 import com.bwsw.tstreams.agents.consumer.Offset.Oldest
 import com.bwsw.tstreams.agents.producer._
